moved package content of edu.ie3.datamodel.dataconnection.* to edu.ie3.datamodel.io.* as these packages have the same intention but io is shorter
