--- src/main/java/edu/ie3/datamodel/dataconnection/source/TypeSource.java
@@ -1,10 +1,0 @@
-/*
- * © 2020. TU Dortmund University,
- * Institute of Energy Systems, Energy Efficiency and Energy Economics,
- * Research group Distribution grid planning and operation
-*/
-package edu.ie3.datamodel.dataconnection.source;
-
-public interface TypeSource extends DataSource {
-  // TODO
-}
