+++ src/main/java/edu/ie3/datamodel/io/sink/DataSink.java
@@ -3,9 +3,9 @@
  * Institute of Energy Systems, Energy Efficiency and Energy Economics,
  * Research group Distribution grid planning and operation
 */
-package edu.ie3.datamodel.dataconnection.sink;
+package edu.ie3.datamodel.io.sink;
 
-import edu.ie3.datamodel.dataconnection.dataconnectors.DataConnector;
+import edu.ie3.datamodel.io.connectors.DataConnector;
 
 /** Describes a class that manages data persistence */
 public interface DataSink {
