+++ src/main/java/edu/ie3/datamodel/io/source/DataSource.java
@@ -3,9 +3,9 @@
  * Institute of Energy Systems, Energy Efficiency and Energy Economics,
  * Research group Distribution grid planning and operation
 */
-package edu.ie3.datamodel.dataconnection.source;
+package edu.ie3.datamodel.io.source;
 
-import edu.ie3.datamodel.dataconnection.dataconnectors.DataConnector;
+import edu.ie3.datamodel.io.connectors.DataConnector;
 
 /** Describes a class that fetches data from a persistence location */
 public interface DataSource {
