+++ src/main/java/edu/ie3/datamodel/io/source/WeatherSource.java
@@ -3,7 +3,7 @@
  * Institute of Energy Systems, Energy Efficiency and Energy Economics,
  * Research group Distribution grid planning and operation
 */
-package edu.ie3.datamodel.dataconnection.source;
+package edu.ie3.datamodel.io.source;
 
 import com.vividsolutions.jts.geom.Point;
 import edu.ie3.datamodel.models.timeseries.IndividualTimeSeries;
